fix(demucs): segment must be int for demucs CLI
The demucs CLI --segment flag does not accept float values. Revert the
segment param type from float to int across config, settings, and the
SeparateConfig model, and update tests accordingly.

+++ config.py
@@ -71,7 +71,7 @@ class Settings(BaseSettings):
     demucs_device: str = "cuda"
     demucs_output_format: str = "wav"
     demucs_mp3_bitrate: int = 320
-    demucs_segment: float | None = None
+    demucs_segment: int | None = None
     demucs_shifts: int | None = None
     demucs_jobs: int | None = None
     demucs_overlap: float | None = None
